Refactor single app file into ddd architecture
