+++ src/users/exceptions.py
@@ -1,0 +1,2 @@
+class UserNotFoundException(Exception):
+    pass
